Fix more RTTI mismatches
Summary:
According to Eric Griffith in private messages, as long as all virtual functions of a class are defined in the header, it's okay to have different compilation units built with different RTTI settings. Since Eric's original fix, a new virtual function was introduced which broke that mix & match, and that's what this diff addresses.

Just because it works it by no means imply that mixing rtti settings is a reasonable thing to do, but that pattern is isolated to a specific shell app. This diff here doesn't change behavior of any shell apps, and therefore it's a safe way to unblock things for the time being.

Reviewed By: EricGriffith, corporateshark

Differential Revision: D59613141

fbshipit-source-id: 6b087f30d9574dfe3af60a2088b6624bcc365593