Add a flag to optionally disable the suballocation feature in the ShaderUniform for Vulkan
Summary: The Suballocation feature was introduced in D42029928 to address the issue when multiple shapes share the same material. The downside of this feature is that it increases the memory usage for all uniform blocks to 64k. This feature is specific to the AR engine Material use case. However, as the IGL and IGLU libraries become open-sourced and are used by more and more projects other than the AR engine, not all use case need this suballocation feature. In this diff, we introduce a setting: tryEnableSuballocation. Suballocation will be enabled for Vulkan only when the flag is set to true (default). It shouldn't have any impact to the current code. But users have additional option to set it to false to disable this feature.

Reviewed By: pixelperfect3, AmesingFlank

Differential Revision: D59694244

fbshipit-source-id: 12798cabe00078a97c5afb4671552af9fe8cab2d