IGL: Change when shader validation happens
Summary: This diff updates when OpenGL shader validation happens. With this diff, it now happens just before drawing rather than at binding time. This allows time for all textures/samplers to be bound. This fixes issues with using it with iOS where textures would not be bound when it was called.

Reviewed By: corporateshark

Differential Revision: D60735118

fbshipit-source-id: 88f63fc99b6e45bf1a20143da3a4572d12f8a99b