igl | Remove `setStencilReferenceValues()`
Summary:
Removed unused code.

Separate back/front stencil reference values are not used on any of our mobile devices and will not be used due no perf reasons.

We are moving towards simplifying the depth-stencil state so that later the stencil state can be merged into PSOs. This is just a refactoring to reduce the "surface area".

Reviewed By: trivedivivek, florihupf

Differential Revision: D60851241

fbshipit-source-id: b9281447970e48cadc9ae5b27194e33d34bd01d2