Allows multiple blocks aliased to the same binding point
Summary:
## This stack

This stack addresses an issue with "Forwarded Uniform Blocks"  which only happens on IOS + OpenGL ES. Specifically, when there is a forwarded uniform block, we're not populating this IGL desc `uniformBlockBindingMap` with binding points of the forwarded block.

(On Android, even if we don't populate that IGL desc, the driver chooses a binding point for those forwarded blocks, which is picked up by the IGL renderpipeline reflection, so we don't observe any issues. But for OpenGL ES on IOS, this fails)

In XOR, the `uniformBlockBindingMap` is populated using this XOR descriptor. We always only populate information for the original block, and ignores any potential forwarding.

The solution is to update IGL's `RenderPipelineDesc` so that it explicitly allows multiple blocks aliased to the same binding point. And then in XOR, we will populate the `uniformBlockBindingMap` conservatively -- adding in all potentially forwarded uniform block names. (This is needed because by the time we build the RenderPipelineDesc, we'd already lost the information of which forwarded blocks there are, which is only available in the SparkSL-generated reflection)

## this diff

Allows multiple blocks aliased to the same binding point

Reviewed By: ChristianK275, corporateshark

Differential Revision: D60920756

fbshipit-source-id: 9f8dac5eccbb38d69dfe14ce557795eb602565fa