Propogate fence wait timeout to RenderCommandEncoder + Propagate result from RenderCommandEncoder
Summary:
# Motivation

I'm adding a way to specify a timeout value for Vulkan's fence wait, and if there's a timeout, we can catch it on the front end and handle it subsequently

# This Diff

So, let's modify the `VulkanContextConfig` struct, which helps to minimize code change on other igl libraries, to accept a property `fenceTimeoutNanoseconds` which will be propagated to `VulkanImmediateCommands::wait`.

Additionally, I'm also propagating the error information from `VulkanImmediateCommands::wait` all the way to `CommandBuffer::createRenderCommandEncoder` via `VulkanContext::checkAndUpdateDescriptorSets`.

Reviewed By: pixelperfect3, corporateshark

Differential Revision: D60803163

fbshipit-source-id: 18fd1c6836500c71acaa83d287dadf3b9602ae52