Do not forever busy wait for command encoder
Summary:
# Do not busy wait forever

While doing my own testing, I realised that there's a busy wait in `VulkanImmediateCommands::acquire`. This may lead to a frozen app if, for some reason, the app could not acquire a command buffer to do a particular rendering.

This diff modifies the behavior of `CommandQueue::createCommandBuffer` that can now return a runtime failure if we could not create a `CommandBuffer` because we could not obtain a `CommandBufferWrapper`, rather than waiting forever.

We then modify`VulkanImmediateCommands::acquire` to return a "empty" `CommandBufferWrapper` if we could not acquire a command buffer after finite attempts.

Reviewed By: pixelperfect3, corporateshark

Differential Revision: D61163381

fbshipit-source-id: 855cdd92fa56202c65330b2311ea55b1ca68f71d