igl | vulkan | Don't switch to a new descriptor pool immediately
Summary:
When creating a new `DescriptorPoolsArena`, we don't have a valid `SubmitHandle` available yet. Therefore, we should delay the actual construction of the Vulkan descriptor pool until it is actually requested and a `SubmitHandle` becomes known.

This will prevent validation errors when trying to reclaim the very first descriptor pool in `DescriptorPoolsArena`.

Reviewed By: EricGriffith, mmaurer

Differential Revision: D61491268

fbshipit-source-id: dd7bb6ee43c17e7b551fa11a1556b413302ad0b5