Fix macro issues with IGLLogLevel
Summary:
# Background

Looks like the fix implemented in D59446783 did not help in preventing a compilation error when compiling an app (see comments).

The compiler still treats `LOG_INFO` as a macro defined outside the `Log.h` file, even though we have them undefined within `Log.h`.

I suspect that when `IGL_LOG_INFO` was used where `LOG_INFO` was previously defined as an Android macro, the compiler would still treat `LOG_INFO` as a macro and replaced it as `6` when doing macro translations.

# The Solution
Simply rename the fields `IGLLogLevel::LOG_INFO` to `IGLLogLevel::IGLLogInfo`, and same for the other enum values.

Reviewed By: EricGriffith, rshest, corporateshark

Differential Revision: D61502653

fbshipit-source-id: da57e673801ac05304c545598797406a0153a9b0