IGL: Add an HDR flag to TextureFormatProperties
Summary:
This diff adds an HDR flag to TextureFormatProperties. This flag is for color textures that use more that 8 bits per component.

The intent of this flag is to help situations where linear <-> sRGB conversions are being done manually. Specifically, it can help with detecting when a particular texture or framebuffer attachment might NOT require manual conversion.

Reviewed By: corporateshark, francoiscoulombe

Differential Revision: D61641678

fbshipit-source-id: c34c7e9e6e85f3351408b92572dd5cc9150211df