Populate glInternalFormat when creating OpenGL texture as externally managed
Summary:
If creating `TextureBuffer` via `createTextureBufferExternal`, the `glInternalFormat_` wasn't correctly populated (remaining at 0), violating the internal consistency of the corresponding Texture object.

When tried to be used later by a client with OpenGL ES backend, this would assert in `getGLInternalTextureFormat` because of that.

This diff correspondingly makes sure that if we use ` createTextureBufferExternal` then `glInternalFormat_` is set to whatever value corresponds to that externally managed texture.

Differential Revision: D61905498