Populate glInternalFormat when creating OpenGL texture as externally managed (#171)
Summary:
Pull Request resolved: https://github.com/facebook/igl/pull/171

If creating `TextureBuffer` via `createTextureBufferExternal`, the `glInternalFormat_` wasn't correctly populated (remaining at 0), violating the internal consistency of the corresponding Texture object.

When tried to be used later by a client with OpenGL ES backend, this would assert in `getGLInternalTextureFormat` because of that.

This diff correspondingly makes sure that if we use ` createTextureBufferExternal` then `glInternalFormat_` is set to whatever value corresponds to that externally managed texture.

Reviewed By: christophpurrer

Differential Revision: D61905498

fbshipit-source-id: 44588677a0168054371382a6ca9ce4528158b8dd