RL Wheels enable scrolling on XRU.
Summary:
This diff enables scrolling on XRU in the RL Wheels validation app. It includes changes to the MediaCarousel.cpp file to signal XRU context to reset when the media is going to change.

It also includes changes to the XruCommandResourceManager.cpp file to adjust vertex state allocation, if more than half vertex state slots are already used. This avoids overflowing XRU HW allocated vertex slots, if more than 16 draws are issued.

Reviewed By: corporateshark

Differential Revision: D61879267

fbshipit-source-id: 69351e54f4ebd512ab1c8aaff8a1987c710c047b