igl | vulkan | tests | Simplify test names so that they can run on the Mac
Summary:
Some tests names were too long (due to the permutation + number of parameters) and that was causing issues on the Mac.

This diff simplifies the names and removes the parameter names. It still provides a customized name for the test to make it easier to debug issues, but now it uses "_" as the separator between the values and no parameter names

Reviewed By: corporateshark, francoiscoulombe

Differential Revision: D62242051

fbshipit-source-id: 3508c164b9a5be9a2038bb2ad0f075048a55805a