Fix bug in deferredTask creation: Use SubmitHandle from current commandbuffer
Summary:
For deferredTask creation if there was a command buffer which was being recorded, we were using the incorrect SubmitHandle from the last submit. This can cause issues since we could destroy resources which are currently in use.

Use the handle from the current command buffer if one exists, otherwise use the last submitted handle

// fb-only Similar issue as D61424399

Reviewed By: corporateshark, AmesingFlank

Differential Revision: D62645168

fbshipit-source-id: 7c716c8dece7cb3788e63c6e6ce0ad206b557948