IGL: Add RenderSessionConfig to use with IRenderSessionFactory
Summary:
This diff adds RenderSessionConfig, which encapsulates some basic properties for setting up a RenderSession. A new helper method is added to IRenderSessionFactory, which accepts a vector of suggested configurations for sessions and lets the factory return the vector of configurations it is requesting to use.

The Windows sample apps are updated to use this.

Reviewed By: rameshviswanathan

Differential Revision: D62716510

fbshipit-source-id: f46e78053601d51cf72521fb2b462f00c6d662cb