IGL: Add IRenderSessionFactory
Summary:
This diff adds IRenderSessionFactory as the new interface for shell implementations to create render sessions. A new, global, createDefaultRenderSessionFactory method is added to construct the factory. In this diff, this is implemented in the .cpp with factory that calls the existing createDefaultRenderSession method. This allows incremental migration to the new pattern.

In later diffs, additional methods will be added to IRenderSessionFactory giving shell apps more control over that the shell creates for them.

The windows sample apps are updated to use this.

Reviewed By: rameshviswanathan

Differential Revision: D62711549

fbshipit-source-id: c6f35cab2fdef08c17915e40032c21bcc8ac7194