IGL: Fix Texture::getAlignment issue when not getting the alignment for the full image range
Summary: This diff fixes the Texture::getAlignment method. It currently assumes stride is for the full image at the specified mip level. However, callers may be trying to get the alignment for a subrange of the image at that miplevel. This diff fixes the issue by allowing both the mip level and the mip level width to be passed in. For the base image (full width, mip level 0), the params can be left as 0. Existing callsites are updated.

Reviewed By: rudybear, corporateshark

Differential Revision: D62688438

fbshipit-source-id: ca897a2d43fe195f7604d9f9ded1f64dc04f88c6