Fix buffer pooling
Summary:
See
https://fb.workplace.com/groups/1381720985696937/posts/1745695652632800 for
some background.

This code doesn't iterate through the whole pool because the success condition
is in the wrong place, so depending on ordering it may leak memory instead.

The code was also not correctly returning the memory after use. This diff fixes
both issues.

This code should likely be refactoring into a separate class with a cleaner
interface, so these bugs are less likely to happen.

Reviewed By: corporateshark, mmaurer

Differential Revision: D62971672

fbshipit-source-id: dc3ec0d5342e2cdf0d89dd263e34e7c3c467187c